refactor(ISCC): 合并擂台赛监控到 ISCCSubmit 并修复若干问题
review 后将独立的 ISCCMonitor 模块合并回 ISCCSubmit，避免同一 ISCC
平台两个模块各自维护 session、HTTP 客户端、DB。改动要点：

- 把 ArenaSolve/TeamArenaSnapshot 和 fetch_team_arena_snapshot 补到
  ISCCSubmit 现有的 ISCCClient 上，统一 session 生命周期
- 在 DataManager 增加 iscc_monitor_target / iscc_monitor_submit 两张表；
  target 带 baseline_ready_at 字段，首轮抓取只建立基线不通知，
  落实 README 里宣称的行为；record_monitor_submit 的 solved_at
  改为仅在原值为空时才覆盖，防止 ISCC 页面相对时间把历史刷掉
- 心跳处理器在现有保活/每日刷新之后追加一次擂台赛监控，使用
  OWNER_ID 对应账号的 session 作为抓取凭据；_last_run_at 改到
  整轮跑完后再更新，修掉原 PR 节流提前更新导致长任务阻塞的 bug
- 新增 isccm添加 / isccm删除 / isccm列表 / isccm检测 四个命令，
  沿用现有 iscc 开关，避免一个平台两个模块开关
- 删除独立的 app/modules/ISCCMonitor 目录
- Dockerfile 安装 tzdata 并 link /etc/localtime，让 TZ=Asia/Shanghai
  在 slim 镜像里真正生效
