+++ app/modules/ISCCSubmit/handlers/handle_message_private.py
@@ -12,12 +12,17 @@ from .. import (
     FLAG_PATTERN,
     HELP_COMMAND,
     MODULE_NAME,
+    MONITOR_ADD_COMMAND,
+    MONITOR_CHECK_COMMAND,
+    MONITOR_LIST_COMMAND,
+    MONITOR_REMOVE_COMMAND,
     NONCE_COMMAND,
     SESSION_COMMAND,
     SWITCH_NAME,
 )
 from .data_manager import DataManager
 from .iscc_client import ISCCClient, ISCCClientError, SubmitResult
+from .monitor_service import MonitorLock, run_monitor_once
 
 
 class PrivateMessageHandler:
@@ -50,6 +55,18 @@ class PrivateMessageHandler:
                 return
             if self.raw_message.lower() == NONCE_COMMAND.lower():
                 await self._handle_query_nonce()
+                return
+            if self.raw_message == MONITOR_LIST_COMMAND:
+                await self._handle_monitor_list()
+                return
+            if self.raw_message.startswith(MONITOR_ADD_COMMAND):
+                await self._handle_monitor_add()
+                return
+            if self.raw_message.startswith(MONITOR_REMOVE_COMMAND):
+                await self._handle_monitor_remove()
+                return
+            if self.raw_message == MONITOR_CHECK_COMMAND:
+                await self._handle_monitor_check()
                 return
             if re.fullmatch(FLAG_PATTERN, self.raw_message):
                 await self._handle_submit_flag()
@@ -222,14 +239,119 @@ class PrivateMessageHandler:
 
     def _help_text(self) -> str:
         return (
-            "ISCC 自动提交帮助\n"
+            "ISCC 自动提交与擂台赛监控帮助\n"
             "iscc：系统管理员开关模块\n"
             "iscc配置 <账号> <密码>：登录并保存账号、密码、session\n"
             "ISCC{xxxxx}：提交 flag 到练武题和擂台题所有未解题目\n"
             f"{SESSION_COMMAND}：查询当前 ISCC session\n"
             f"{NONCE_COMMAND}：查询当前练武题和擂台题 nonce\n"
-            f"{SWITCH_NAME}{MENU_COMMAND}：查看模块菜单"
+            f"{MONITOR_ADD_COMMAND} <team_id> [备注]：添加擂台赛监控目标\n"
+            f"{MONITOR_REMOVE_COMMAND} <team_id>：删除擂台赛监控目标\n"
+            f"{MONITOR_LIST_COMMAND}：查看擂台赛监控目标列表\n"
+            f"{MONITOR_CHECK_COMMAND}：立即触发一次擂台赛监控轮询\n"
+            f"{SWITCH_NAME}{MENU_COMMAND}：查看模块菜单\n"
+            "说明：开启模块并配置好账号后，心跳会驱动 session 保活/每日刷新；\n"
+            "同时会按节流间隔轮询所有监控目标的擂台赛解题详情，新通过题目会私聊通知管理员。"
         )
+
+    # ==================== 擂台赛监控命令 ====================
+
+    async def _handle_monitor_list(self):
+        with DataManager() as dm:
+            targets = dm.list_monitor_targets()
+            account = dm.get_account(self.user_id)
+
+        if not targets:
+            await self._reply(
+                "当前没有监控任何擂台赛 team id。\n"
+                "可以通过 `isccm添加 <team_id> [备注]` 添加监控对象。"
+            )
+            return
+
+        lines = [f"当前监控中的擂台赛 team id 共 {len(targets)} 个："]
+        for idx, t in enumerate(targets, start=1):
+            remark = t.get("remark") or ""
+            team_name = t.get("team_name") or ""
+            total_score = t.get("total_score") or ""
+            baseline = "基线已建立" if t.get("baseline_ready_at") else "基线待建立"
+            line = (
+                f"{idx}. {t['team_id']}"
+                + (f" [{remark}]" if remark else "")
+                + (f" 队名：{team_name}" if team_name else "")
+                + (f" {total_score}" if total_score else "")
+                + f"（{baseline}）"
+            )
+            lines.append(line)
+        lines.append("")
+        lines.append(
+            "抓取凭据："
+            + (
+                "使用当前 ISCC 账号的 session"
+                if account and account.get("session")
+                else "当前账号未登录，请先 `iscc配置` 保存账号后再监控"
+            )
+        )
+        await self._reply("\n".join(lines))
+
+    async def _handle_monitor_add(self):
+        parts = self.raw_message.split(maxsplit=2)
+        if len(parts) < 2 or not parts[1].strip():
+            await self._reply("用法：isccm添加 <team_id> [备注]")
+            return
+        team_id = parts[1].strip()
+        remark = parts[2].strip() if len(parts) == 3 else ""
+
+        with DataManager() as dm:
+            is_new = dm.add_monitor_target(team_id, remark)
+
+        if is_new:
+            await self._reply(
+                f"已添加擂台赛监控：{team_id}"
+                + (f"（备注：{remark}）" if remark else "")
+                + "\n下一次心跳会尝试拉取该队伍的擂台赛详情并建立基线。"
+            )
+        else:
+            await self._reply(
+                f"该 team id 已在监控中，已更新备注为：{remark or '(空)'}"
+            )
+
+    async def _handle_monitor_remove(self):
+        parts = self.raw_message.split(maxsplit=1)
+        if len(parts) < 2 or not parts[1].strip():
+            await self._reply("用法：isccm删除 <team_id>")
+            return
+        team_id = parts[1].strip()
+        with DataManager() as dm:
+            ok = dm.remove_monitor_target(team_id)
+        if ok:
+            await self._reply(
+                f"已删除擂台赛监控：{team_id}，相关历史提交记录已清除。"
+            )
+        else:
+            await self._reply(f"未找到擂台赛监控 team id：{team_id}")
+
+    async def _handle_monitor_check(self):
+        lock = MonitorLock.get()
+        if lock.locked():
+            await self._reply("已有一次擂台赛监控轮询在执行，请稍后再试。")
+            return
+
+        with DataManager() as dm:
+            account = dm.get_account(self.user_id)
+        if not account or not account.get("session"):
+            await self._reply(
+                "当前账号尚未登录或 session 为空，请先发送：iscc配置 <账号> <密码>"
+            )
+            return
+
+        await self._reply("正在立即执行一次 ISCC 擂台赛监控轮询...")
+        async with lock:
+            await run_monitor_once(
+                self.websocket,
+                self.user_id,
+                account=account,
+                manual_trigger=True,
+            )
 
     async def _reply(self, text: str, auto_delete: bool = False):
         note = "del_msg=30" if auto_delete else ""
